requirements.txt
Flask
python-dotenv
openai
PyMuPDF
Werkzeug
gunicorn
pytesseract
Pillow
